requirements.txt
pandas==2.2.2
numpy==1.26.4
requests==2.32.3
PyYAML==6.0.2
jinja2==3.1.4
python-dateutil==2.9.0.post0
matplotlib==3.8.4
